requirements.txt
hikari
python-dotenv
hikari-lightbulb
